Layouts for vector and matrix chapters separated, populated temp page for vector-ops
